+++ src/espeech/multispeech_voices.hpp
@@ -38,7 +38,7 @@ public:
 
 private:
   // Reference values:
-  static const double ref_freq = 16000.0, rate_scale = 200.0;
+  static const float ref_freq = 16000.0, rate_scale = 200.0;
 
   // Patterns for parameters extraction:
   const boost::regex freq_extractor, pitch_extractor, rate_extractor, volume_extractor,
